-add the missing link between frontend and backend, chat now works, result is three microservices for now, due to time management, I think this is short-term easiest solution
