+++ config.toml
@@ -1,5 +1,5 @@
 [llm]
 client_type = "ollama"
-model_type = "llama3.1"
+model_type = "qwen2.5:7b"
 max_tokens = 1000
 host = "http://127.0.0.1:11434"
