+++ pyproject.toml
@@ -86,7 +86,9 @@ nodejs = ">=22.13.0,<22.14"
 itsdangerous = ">=2.2.0,<3"
 
 [tool.pixi.tasks]
-mcp_server="PYTHONPATH=src python -m studio.main"
+backend="docker compose up"
+mcp_server="PYTHONPATH=src/llm-worker python src/llm-worker/mcp_servers/main.py"
+llm_worker="PYTHONPATH=src python src/llm-worker/main.py"
 vscode = "code ."
 lint = "ruff check"
 typecheck = "mypy ."
