+++ src/backend/__init__.py
@@ -1,0 +1,3 @@
+from .main import app
+
+# This file just exposes the FastAPI instance for uvicorn
