--- src/backend/endpoints/__init__.py
@@ -1,3 +1,0 @@
-from .endpoints import app
-
-# This file just exposes the FastAPI instance for uvicorn
