+++ src/llm-worker/lib/models.py
