+++ src/llm-worker/lib/schemagenerators.py
