+++ src/llm-worker/lib/tools.py
